+++ tests/FSharp.Data.GraphQL.IntegrationTests.Server/Program.fs
@@ -1,4 +1,4 @@
-namespace FSharp.Data.GraphQL.Samples.GiraffeServer
+﻿namespace FSharp.Data.GraphQL.IntegrationTests.Server
 
 open Microsoft.AspNetCore
 open Microsoft.AspNetCore.Hosting