+++ tests/FSharp.Data.GraphQL.IntegrationTests/RemoteProviderTests.fs
@@ -8,11 +8,6 @@ open FSharp.Data.GraphQL
 type Provider = GraphQLProvider<"http://localhost:8084">
 type Episode = Provider.Types.Episode
 
-// We should be able to create instances of schema types.
-let ball = Provider.Types.Ball(form = "Spheric", format = "Spheric", id = "1", order = 0, size = 1.11)
-let box = Provider.Types.Box(form = "Cubic", format = "Cubic", id = "2", order = 1, size = 2.0)
-let things : Provider.Types.IThing list = [ball; box]
-
 module SimpleOperation =
     let operation = 
         Provider.Operation<"""query Q {
@@ -36,9 +31,8 @@ module SimpleOperation =
     type Operation = Provider.Operations.Q
     
     let validateResult (result : Operation.OperationResult) =
-        result.CustomData.IsSome |> equals true
-        result.CustomData.Value.ContainsKey("documentId") |> equals true
-        result.Errors |> equals None
+        result.CustomData.ContainsKey("documentId") |> equals true
+        result.Errors |> equals [||]
         result.Data.IsSome |> equals true
         result.Data.Value.Hero.IsSome |> equals true
         result.Data.Value.Hero.Value.AppearsIn |> equals [| Episode.NewHope; Episode.Empire; Episode.Jedi |]
@@ -66,21 +60,6 @@ module SimpleOperation =
         actual |> equals expected
 
 [<Fact>]
-let ``Should be able to pretty print schema types`` () =
-    let actual = normalize <| sprintf "%A" things
-    let expected = normalize <| """[{Form = "Spheric";
-      Format = "Spheric";
-      Id = "1";
-      Order = 0;
-      Size = 1.11;};
-      {Form = "Cubic";
-      Format = "Cubic";
-      Id = "2";
-      Order = 1;
-      Size = 2.0;}]"""
-    actual |> equals expected
-
-[<Fact>]
 let ``Should be able to start a simple query operation synchronously`` () =
     SimpleOperation.operation.Run()
     |> SimpleOperation.validateResult
@@ -90,26 +69,6 @@ let ``Should be able to start a simple query operation asynchronously`` () =
     SimpleOperation.operation.AsyncRun()
     |> Async.RunSynchronously
     |> SimpleOperation.validateResult
-
-[<Fact>]
-let ``Should be able to start a simple query operation synchronously with custom HTTP headers`` () =
-    let userData = Guid.NewGuid().ToString()
-    let context = Provider.GetContext([| "UserData", userData |], "http://localhost:8084")
-    let result = SimpleOperation.operation.Run(context)
-    SimpleOperation.validateResult result
-    result.CustomData.IsSome |> equals true
-    result.CustomData.Value.ContainsKey("userData") |> equals true
-    result.CustomData.Value.["userData"] |> equals (upcast userData)
-
-[<Fact>]
-let ``Should be able to start a simple query operation asynchronously with custom HTTP headers`` () =
-    let userData = Guid.NewGuid().ToString()
-    let context = Provider.GetContext([| "UserData", userData |], "http://localhost:8084")
-    let result = SimpleOperation.operation.AsyncRun(context) |> Async.RunSynchronously
-    SimpleOperation.validateResult result
-    result.CustomData.IsSome |> equals true
-    result.CustomData.Value.ContainsKey("userData") |> equals true
-    result.CustomData.Value.["userData"] |> equals (upcast userData)
 
 [<Fact>]
 let ``Should be able to use pattern matching methods on an union type`` () =
@@ -148,79 +107,6 @@ let ``Should be able to use pattern matching methods on an union type`` () =
         SimpleOperation.Operation.Types.Hero.Friends.Droid(name = "C-3PO", primaryFunction = "Protocol")
         SimpleOperation.Operation.Types.Hero.Friends.Droid(name = "R2-D2", primaryFunction = "Astromech") |]
   
-module InterfaceOperation =
-    let operation =
-        Provider.Operation<"""query TestQuery {
-            things {
-              id
-              format
-              ...on Ball {
-                form
-              }
-              ...on Box {
-                form
-              }
-            }
-          }""">()
-    
-    type Operation = Provider.Operations.TestQuery
-
-    let validateResult (result : Operation.OperationResult) =
-        result.CustomData.IsSome |> equals true
-        result.CustomData.Value.ContainsKey("documentId") |> equals true
-        result.Errors |> equals None
-        result.Data.IsSome |> equals true
-        let expectedThings : Operation.Types.Things.Thing [] =
-          [| Operation.Types.Things.Ball(id = "1", format = "Spheric", form = "Spheric")
-             Operation.Types.Things.Box(id = "2", format = "Cubic", form = "Cubic") |]
-        result.Data.Value.Things |> equals expectedThings
-        let actual = normalize <| sprintf "%A" result.Data
-        let expected = normalize <| """Some
-            {Things = [|{Form = "Spheric";
-            Format = "Spheric";
-            Id = "1";};
-            {Form = "Cubic";
-            Format = "Cubic";
-            Id = "2";}|];}"""
-        actual |> equals expected
-
-[<Fact>]
-let ``Should be able to run a query with interface types synchronously`` () =
-    InterfaceOperation.operation.Run()
-    |> InterfaceOperation.validateResult
-
-[<Fact>]
-let ``Should be able to run a query with interface types asynchronously`` () =
-    InterfaceOperation.operation.AsyncRun()
-    |> Async.RunSynchronously
-    |> InterfaceOperation.validateResult
-
-[<Fact>]
-let ``Should be able to use pattern matching methods on an interface type`` () =
-    let result = InterfaceOperation.operation.Run()
-    result.Data.IsSome |> equals true
-    let things = result.Data.Value.Things
-    let balls = things |> Array.choose (fun x -> x.TryAsBall())
-    balls |> equals [| InterfaceOperation.Operation.Types.Things.Ball(id = "1", format = "Spheric", form = "Spheric") |]
-    let boxes = things |> Array.choose (fun x -> x.TryAsBox())
-    boxes |> equals [| InterfaceOperation.Operation.Types.Things.Box(id = "2", format = "Cubic", form = "Cubic") |]
-    try
-      things |> Array.map (fun x -> x.AsBall()) |> ignore
-      failwith "Expected exception when trying to get all things as balls!"
-    with _ -> ()
-    try
-      things |> Array.map (fun x -> x.AsBox()) |> ignore
-      failwith "Expected exception when trying to get all things as boxes!"
-    with _ -> ()
-    things
-    |> Array.filter (fun x -> x.IsBall())
-    |> Array.map (fun x -> x.AsBall())
-    |> equals [| InterfaceOperation.Operation.Types.Things.Ball(id = "1", format = "Spheric", form = "Spheric") |]
-    things
-    |> Array.filter (fun x -> x.IsBox())
-    |> Array.map (fun x -> x.AsBox())
-    |> equals [| InterfaceOperation.Operation.Types.Things.Box(id = "2", format = "Cubic", form = "Cubic") |]
-
 module MutationOperation =
     let operation =
         Provider.Operation<"""mutation M {
@@ -234,9 +120,8 @@ module MutationOperation =
     type Operation = Provider.Operations.M
 
     let validateResult (result : Operation.OperationResult) =
-        result.CustomData.IsSome |> equals true
-        result.CustomData.Value.ContainsKey("documentId") |> equals true
-        result.Errors |> equals None
+        result.CustomData.ContainsKey("documentId") |> equals true
+        result.Errors |> equals [||]
         result.Data.IsSome |> equals true
         result.Data.Value.SetMoon.IsSome |> equals true
         result.Data.Value.SetMoon.Value.Id |> equals "1"
@@ -254,34 +139,13 @@ let ``Should be able to run a mutation asynchronously`` () =
     |> Async.RunSynchronously
     |> MutationOperation.validateResult
 
-module VariablesOperation =
-    let operation =
-        Provider.Operation<"""query Q2($filter: ThingFilter!) {
-            things(filter: $filter) {
-              id
-              format
-            }
-          }""">()
-
-    type Operation = Provider.Operations.Q2
-
-    let validateResult (filter : Provider.Types.ThingFilter) (result : Operation.OperationResult) =
-        result.CustomData.IsSome |> equals true
-        result.CustomData.Value.ContainsKey("documentId") |> equals true
-        result.Errors |> equals None
-        result.Data.IsSome |> equals true
-        hasItems result.Data.Value.Things
-        result.Data.Value.Things |> Array.forall (fun x -> x.Format = filter.Format) |> equals true
-
-
 module FileOperation =
     let fileOp = Provider.Operation<"operation.graphql">()
     type Operation = Provider.Operations.FileOp
     
     let validateResult (result : Operation.OperationResult) =
-        result.CustomData.IsSome |> equals true
-        result.CustomData.Value.ContainsKey("documentId") |> equals true
-        result.Errors |> equals None
+        result.CustomData.ContainsKey("documentId") |> equals true
+        result.Errors |> equals [||]
         result.Data.IsSome |> equals true
         result.Data.Value.Hero.IsSome |> equals true
         result.Data.Value.Hero.Value.AppearsIn |> equals [| Episode.NewHope; Episode.Empire; Episode.Jedi |]
@@ -307,20 +171,6 @@ module FileOperation =
             HomePlanet = Some "Tatooine";
             Name = Some "Luke Skywalker";};}"""
         actual |> equals expected
-
-
-[<Fact>]
-let ``Should be able to run a query with variables syncrhonously`` () =
-    let filter = Provider.Types.ThingFilter(format = "Cubic")
-    VariablesOperation.operation.Run(filter)
-    |> VariablesOperation.validateResult filter
-
-[<Fact>]
-let ``Should be able to run a query with variables asyncrhonously`` () =
-    let filter = Provider.Types.ThingFilter(format = "Cubic")
-    VariablesOperation.operation.AsyncRun(filter)
-    |> Async.RunSynchronously
-    |> VariablesOperation.validateResult filter
 
 [<Fact>]
 let ``Should be able to run a query from a query file`` () =